% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherFunction.R
\name{ASmapIso}
\alias{ASmapIso}
\title{AS mapped to isoforms}
\usage{
ASmapIso(object, gtf, AStype = "exonic", cores = 10)
}
\arguments{
\item{object}{A FAScore object. see the constructor function '\code{\link{FAScoreDataSet}}'.}

\item{gtf}{An annotation file was consistent with identifying AS.}

\item{AStype}{The types of AS event sites, including 'exonic' (exonic sites in genomes) and 'intronic' (intronic sites in genomes, eg. splicing sits).}

\item{cores}{The number of threads. (default:10)}
}
\value{
Return a GRange object. The alternative splicing sites were aligned to transcripts by an annotation file.
}
\description{
AS mapped to isoforms
}
