% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateDyScore.R
\name{CalcuDyScore}
\alias{CalcuDyScore}
\title{Calculate the dynamic score}
\usage{
CalcuDyScore(object, maxRange, maxSlope = 1, type = "Gene")
}
\arguments{
\item{object}{A FAScore object. see the constructor function '\code{\link{FAScoreDataSet}}'.}

\item{maxRange}{A numeric vector. You can set the maximal value which would be scaled to 1 of range values of all features.}

\item{maxSlope}{A numeric vector. You can set the maximal value which would be scaled to 1 of slope values of all features. (default is 1)}

\item{type}{A character. Feature types, such as "Gene" and "AS".}
}
\value{
Return a numeric vector. Dynamic score of feature type was stored in 'DyScore' slot of a FAScore object.
}
\description{
Dynamic score was calculated using the scaled sum of the range, slope, Spearman correlation, Spearman p-value, linear fitting slope and
fitting p-value. The ranges value were scaled to 0~1 (0, 0.2, 0.4, 0.6, 0.8, 1) by maximal range value. The slopes were scale to -1~1. The
slopes of no less than 1 were set to 1, and of not greater than -1 were set to -1. The p-value less than 0.05 was set to 1, otherwise set to 0.
}
