% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateFAScore.R
\name{CalcuFAScore}
\alias{CalcuFAScore}
\title{Calculate Functional score using trained random forest model}
\usage{
CalcuFAScore(object, model = NULL)
}
\arguments{
\item{object}{A FAScore object. see the constructor function '\code{\link{FAScoreDataSet}}'.}

\item{model}{A trained model to predict functional AS. By default, the package uses the pre-trained model,
but users can also choose to apply their own trained model.}
}
\value{
Return a FAScore object. The prediction results were added to the 'RFpredict' slot.
}
\description{
Calculate Functional score using trained random forest model
}
