% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateValues.R
\name{CalcuFeature}
\alias{CalcuFeature}
\title{Calculate all features values}
\usage{
CalcuFeature(object, group.by, cores = 10)
}
\arguments{
\item{object}{A FAScore object. see the constructor function '\code{\link{FAScoreDataSet}}'.}

\item{group.by}{Name of one metadata column to group samples by (for example, Stage or CellType).}

\item{cores}{The number of threads. (default is 10)}
}
\value{
Return a FAScore object. The correlation table including the rho and p-value was stored in 'Correlation' slot; tau value was stored in 'Tau' slot; the range
value was stored in 'Range' slot; linear fitting table including the intercept, slope, r2 and p-value was stored in 'Linear' slot of a
FAScore object.
}
\description{
Calculate all features values including the correlation, tau, ranges and linear fitting scores of AS and Gene types. The correlation values of
samples groups with genes expression or AS PSI were calculated by Spearman's rank analysis. A stage or cell type specificity index value, Tau,
valued between 0 for housekeeping and 1 for tissue specific genes or AS, was calculated using the Itai Yanai et al (2005) method. The range value
was defined to the difference of average genes expression or AS PSI of all stages or cell types. The samples groups with genes expression or
AS PSI were fitted using linear method.
}
