% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateFAScore.R
\name{calcuGMM}
\alias{calcuGMM}
\title{classify by GMM}
\usage{
calcuGMM(object, LOG = NULL, silent = F, main = NULL)
}
\arguments{
\item{object}{A FAScore object. see the constructor function '\code{\link{FAScoreDataSet}}'.}

\item{LOG}{A numeric vector. FAScore log transformed, 2 or 10. Default is NULL.}

\item{silent}{A logic value. Drawing the plot.}

\item{main}{The title of the plot.}
}
\value{
Return a FAScore object or a figure. The classification result was added to the 'RFpredict' slot.
And the fitting result of GMM was added to the 'GMM' slot.
}
\description{
classify by GMM
}
