% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateValues.R
\name{call_cor}
\alias{call_cor}
\title{Calculate Spearman's rank correlation of genes or AS}
\usage{
call_cor(mat, meta, group.by = "CellType", exact = NULL, cores = 10)
}
\arguments{
\item{mat}{Expression matrix of genes or PSI matrix of AS.}

\item{meta}{Sample information.}

\item{group.by}{Name of one metadata column to group samples by (for example, Stage or CellType).}

\item{exact}{a logical indicating whether an exact p-value should be computed. Used for Spearman's rho.}

\item{cores}{The number of threads. (default is 10)}
}
\value{
A data frame including correlation estimate value and p value columns.
}
\description{
Calculate Spearman's rank correlation of genes or AS
}
