% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateFAScore.R
\name{ChooseIso}
\alias{ChooseIso}
\title{Choose one transcript by the maximum expression from multiple transcripts}
\usage{
ChooseIso(object, cores = 10)
}
\arguments{
\item{object}{A FAScore object. see the constructor function '\code{\link{FAScoreDataSet}}'.}

\item{cores}{A numeric. The number of threads. (default:10)}
}
\value{
Return a transcript ID from multiple mapped transcripts. Storing in 'rowRanges' slot.
}
\description{
Choose one transcript by the maximum expression from multiple transcripts
}
