% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExDataSet}
\alias{ExDataSet}
\title{Example data sets}
\format{
An object of class \code{list} of length 4.
}
\usage{
ExDataSet
}
\description{
A list of AS, genes, transcripts and sample information matrix.
}
\examples{
data(ExDataSet)

}
\keyword{datasets}
