% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Class.R
\docType{class}
\name{FAScore-class}
\alias{FAScore-class}
\alias{FAScoreDataSet}
\title{define a FAScore S4 class}
\usage{
FAScoreDataSet(AS, colData, GENE, ISOFORM)
}
\arguments{
\item{AS}{A data frame of PSI values of the identified AS.}

\item{colData}{A data frame of samples information including sample group column. The order of rows is the same as the order of columns in 'AS', 'Gene', 'ISOFORM' table.}

\item{GENE}{A data frame of normalized expression of the genes.}

\item{ISOFORM}{A data frame of normalized expression of the transcripts including gene id column.}
}
\value{
Return a S4 object of FAScore class.
}
\description{
The FAScore object is the functional alternative splicing score (FAScore).
It stores all information associated with the dataset, including gene expression, PSI,
isoform expression, sample information, analysis etc. All that is needed to construct
a FAScore object is the gene and isoform expression matrix, alternative splicing (AS)
PSI matrix (rows are AS, columns are samples), and sample information (rows are samples).
}
\section{Slots}{

\describe{
\item{\code{assays}}{A SimpleAssays class. The PSI matrix of alternative splicing is stored in \code{RangedSummarizedExperiment} Container.}

\item{\code{GENE}}{The normalized gene expression are stored in \code{RangedSummarizedExperiment} Container.}

\item{\code{ISOFORM}}{The normalized isoform expression are stored in \code{RangedSummarizedExperiment} Container.}

\item{\code{rowRanges}}{The \code{GRanges} format information of all alternative splicing are stored in \code{RangedSummarizedExperiment} Container.}

\item{\code{colData}}{The sample information are stored in \code{RangedSummarizedExperiment} Container.}

\item{\code{Correlation}}{A list including the Spearman's rank correlation coefficient and p-value of genes and AS events.}

\item{\code{Tau}}{A list including specificity index (τ) value of genes and AS events, which is a quantitative, graded scalar measure of the specificity of an expression profile. The greater the value, the greater the specificity.}

\item{\code{Range}}{A list including range value (R) of genes and AS events.}

\item{\code{Linear}}{A list including linear fitting coefficient (intercept, slop and r2) and p-value of genes and AS events.}

\item{\code{DyScore}}{A list including the dynamic value of genes and AS events.}

\item{\code{Structure}}{The data frame of isoform structural scores from APPRIS database. The columns include the Firestar (AA residues), Matador3D (3D structure), CORSAIR (Conservation), SPADE (Domain), THUMP (trans-membrane helix), SignalP (signal peptide), TargetP (target peptide).}

\item{\code{RFpredict}}{The predicted values using RF modeling storing in FAScore column.}

\item{\code{GMM}}{The result of gaussian mixture model (GMM).}

\item{\code{NAMES}}{The slot is from \code{RangedSummarizedExperiment} Container.}

\item{\code{elementMetadata}}{The slot is from \code{RangedSummarizedExperiment} Container.}

\item{\code{metadata}}{An optional list of arbitrary content describing the overall experiment.}
}}

