% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateValues.R
\name{FitValue}
\alias{FitValue}
\title{fit linear}
\usage{
FitValue(object, group.by, type = "AS", cores = 10)
}
\arguments{
\item{object}{A FAScore object. see the constructor function '\code{\link{FAScoreDataSet}}'.}

\item{group.by}{Name of one metadata column to group samples by (for example, Stage or CellType).}

\item{type}{A character. Feature types, such as "Gene" and "AS".}

\item{cores}{The number of threads. (default is 10)}
}
\value{
A data frame including the intercept, slope, r2 and pvalue columns.
}
\description{
fit linear
}
