% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherFunction.R
\name{HostGene}
\alias{HostGene}
\title{AS mapped to host genes}
\usage{
HostGene(object, gtf, cores = 10)
}
\arguments{
\item{object}{A FAScore object. see the constructor function '\code{\link{FAScoreDataSet}}'.}

\item{gtf}{An annotation file was consistent with identifying AS.}

\item{cores}{The number of threads. (default is 10)}
}
\value{
Return a GRange object. The alternative splicing sites were aligned to genes by an annotation file.
}
\description{
AS mapped to host genes
}
