% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateFAScore.R
\name{matchAppris}
\alias{matchAppris}
\title{Add structure related scores}
\usage{
matchAppris(object, gtf, species = NULL, appris = NULL)
}
\arguments{
\item{object}{A FAScore object. see the constructor function '\code{\link{FAScoreDataSet}}'.}

\item{gtf}{An annotation file was consistent with identifying AS.}

\item{species}{Latin abbreviation for species, including 'HomSap'(GRCh38), 'MacMul'(Mmul10), 'RatNor'(Rnor6.0), 'MusMus'(GRCm38) and 'DanRer'(GRCz11).}

\item{appris}{A data frame. If your species is not in our dataset (species = NULL), you can download the file corresponding species from APPRIS database, and
pass to this parameter. (default is NULL)}
}
\value{
Return a structure score table stored in the 'Structure' slot.
}
\description{
Add structure related scores including the functional residues (Firestar), structure score (Matador), conservation (CORSAIR),
domain score (SPADE), transmembrane helices (THUMP) and signal and target peptides (CRASH) scores by APPRIS database
(\url{https://appris.bioinfo.cnio.es/#/}).
}
