% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateFAScore.R
\name{mergeFeatures}
\alias{mergeFeatures}
\title{Merge features}
\usage{
mergeFeatures(object)
}
\arguments{
\item{object}{A FAScore object. see the constructor function '\code{\link{FAScoreDataSet}}'.}
}
\value{
a data frame object.
}
\description{
Merge features
}
