% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigurePlot.R
\name{plotDistrib}
\alias{plotDistrib}
\title{Visualization of the GMM result}
\usage{
plotDistrib(model, names = NULL, LOG, xlab = "FAScore", ylab = "Density")
}
\arguments{
\item{model}{A FAScore object. see the constructor function '\code{\link{FAScoreDataSet}}'.}

\item{names}{The title of the plot.}

\item{LOG}{A numeric vector. FAScore log transformed, 2 or 10. Default is NULL.}

\item{xlab}{The title of the x axis.}

\item{ylab}{The title of the y axis.}
}
\value{
ggplot
}
\description{
Visualization of the GMM result
}
