% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigurePlot.R
\name{plotFAScore}
\alias{plotFAScore}
\title{Visualization of functional AS scores}
\usage{
plotFAScore(object, ASID = NULL, label = NULL, color = "red", size = 15)
}
\arguments{
\item{object}{A FAScore object. see the constructor function '\code{\link{FAScoreDataSet}}'.}

\item{ASID}{A list of highlight AS ids. (default is NULL).}

\item{label}{A list of alias of highlight AS. (default is NULL).}

\item{color}{A character vector. The length of highlight label colors is same as the length of label list.}

\item{size}{The size aesthetic control the size of text.}
}
\value{
A ggplot object.
}
\description{
The dot plot of functional AS score with custom labeled AS coloring.
}
