% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FigurePlot.R
\name{plotSmooth}
\alias{plotSmooth}
\title{Visualization of the expression or PSI of Genes and AS}
\usage{
plotSmooth(object, label, size = 15, color = c("#66c2a5", "#fc8d62"))
}
\arguments{
\item{object}{A FAScore object. see the constructor function '\code{\link{FAScoreDataSet}}'.}

\item{label}{Vector of AS ids to plot.}

\item{size}{Base font size, given in pts (default: 15). see the '\code{\link{theme_test}}'.}

\item{color}{Vector of colors corresponding to 'AS' and 'Gene'. Default: c("#66c2a5", "#fc8d62")}
}
\value{
A ggplot object.
}
\description{
The dot plot of the genes of normalized expression and the AS of PSI value for all samples smoothing by '\code{\link{loess}}' function.
}
